+++ src/kardy/base.nim
@@ -28,21 +28,11 @@ type
     deck*: seq[CardId]
     actions*: seq[Action]
 
-
-  CardState* = ref object
-    ## This object holds the state of the card in game
-    id*: CardId
-    probabilities*: seq[Probability]
-    discarded*: bool
-
-  CardsState* = seq[CardState]
-
 using
   id: CardId
   card: CardSetting
   settings: Settings
   state: State
-  cardsState: CardsState
 
 converter cardIdToInt*(id): int =
   int id
@@ -52,11 +42,6 @@ func get*(cards: seq[CardSetting]; id): CardSetting =
   for card in cards:
     if id == card.id:
       return card
-func get*(cardsState; id): CardState =
-  new result
-  for cardState in cardsState:
-    if id == cardState.id:
-      return cardState
 
 func newAction*(id; kind: ActionKind; prob: Probability = 0.0): Action =
   new result
@@ -72,19 +57,11 @@ func discarded*(card; state): bool =
     if action.kind == Discard and action.cardId == card.id:
       return true
 
-func calculateCardsState*(settings, state): CardsState =
-  for cardId in state.deck:
-    var cardState = new CardState
-    cardState.id = cardId
-    result.add cardState
-
+func probabilities*(card; state): seq[Probability] =
   for action in state.actions:
-    let cardState = result.get action.cardId
-    case action.kind:
-      of Discard:
-        cardState.discarded = true
-      of NewProbability:
-        cardState.probabilities.add action.probability
+    if action.kind == NewProbability:
+      if action.cardId == card.id:
+        result.add action.probability
 
 func summary*(probs: seq[Probability]): Probability =
   var sum: float