+++ src/kardy/storage.nim
@@ -31,5 +31,6 @@ proc getState*(default: State): State =
   try:
     result.fromJson window.localStorage.getItem(localStorageStateKey).`$`.parseJson
   except:
+    echo "getCurrentExceptionMsg()"
     echo getCurrentExceptionMsg()
     result = default